requirements.txt
openai
requests
python-dotenv
ipywidgets
gradio
fastapi
unicorn